Conflicts:
	src/ToggleView.cpp
	src/ToggleView.h
	src/dependencies/tinyxml/tinystr.h
	src/dependencies/tinyxml/tinyxml.h
	src/dependencies/zlib/include/ioapi.h
	src/dependencies/zlib/include/unzip.h
	src/formats/FFT.cpp
	src/formats/FFT.h
	src/formats/HOSA.cpp
	src/formats/HOSA.h
	src/formats/MP2k.cpp
	src/formats/MP2k.h
	src/formats/SegSat.cpp
	src/formats/SegSat.h
	src/lib/minizip/ioapi.h
	src/lib/minizip/unzip.h
	src/lib/tinyxml/lib/win32/tinyxmlSTL-mt.lib
	src/lib/tinyxml/lib/win32/tinyxmlSTL-mtd.lib
	src/lib/tinyxml/lib/x64/tinyxmlSTL-mt.lib
	src/lib/tinyxml/lib/x64/tinyxmlSTL-mtd.lib
	src/lib/tinyxml/src/tinystr.h
	src/lib/tinyxml/src/tinyxml.h
	src/lib/tinyxml/tinystr.h
	src/lib/tinyxml/tinyxml.h
	src/lib/zlib/include/ioapi.h
	src/lib/zlib/include/unzip.h
	src/lib/zlib/lib/win32/zlibstat.lib
	src/lib/zlib/lib/x64/zlibstat.lib
	src/main/SeqTrack.cpp
	src/main/SeqTrack.h
	src/main/SeqVoiceAllocator.cpp
	src/main/SeqVoiceAllocator.h
	src/main/ToggleView.cpp
	src/main/ToggleView.h
	src/main/VGMItem.h
	src/main/common.h
	src/main/formats/CapcomSnesSeq.cpp
	src/main/formats/FFT.cpp
	src/main/formats/FFT.h
	src/main/formats/FFTSeq.cpp
	src/main/formats/FFTSeq.h
	src/main/formats/HOSA.cpp
	src/main/formats/HOSA.h
	src/main/formats/HOSAScanner.cpp
	src/main/formats/HOSASeq.cpp
	src/main/formats/HOSASeq.h
	src/main/formats/HeartBeatPS1Seq.cpp
	src/main/formats/KonamiSnesInstr.cpp
	src/main/formats/KonamiSnesScanner.cpp
	src/main/formats/KonamiSnesScanner.h
	src/main/formats/KonamiSnesSeq.cpp
	src/main/formats/MP2k.cpp
	src/main/formats/MP2k.h
	src/main/formats/MP2kSeq.cpp
	src/main/formats/MP2kSeq.h
	src/main/formats/NinSnesFormat.h
	src/main/formats/NinSnesInstr.cpp
	src/main/formats/NinSnesInstr.h
	src/main/formats/NinSnesScanner.cpp
	src/main/formats/NinSnesScanner.h
	src/main/formats/NinSnesSeq.cpp
	src/main/formats/NinSnesSeq.h
	src/main/formats/QSoundSeq.cpp
	src/main/formats/RareSnesSeq.cpp
	src/main/formats/SNESDSP.h
	src/main/formats/SegSat.cpp
	src/main/formats/SegSat.h
	src/main/formats/SegSatSeq.cpp
	src/main/formats/SegSatSeq.h
	src/main/formats/SquSnesFormat.h
	src/main/formats/SquSnesScanner.cpp
	src/main/formats/SquSnesScanner.h
	src/main/formats/SquSnesSeq.h
	src/main/formats/SquareSnesSeq.cpp
	src/main/formats/TaitoF3Format.h
	src/main/formats/TaitoF3Scanner.cpp
	src/main/formats/TaitoF3Scanner.h
	src/main/formats/TriAcePS1Scanner.cpp
	src/ui/windows/ToggleView.cpp
	src/ui/windows/ToggleView.h

+++ src/main/VGMSamp.cpp
@@ -152,5 +152,5 @@ bool VGMSamp::SaveAsWav(const wchar_t* filepath)
 	
 	//DLSWave dlswave(1, channels, rate, rate*blockAlign, blockAlign, bps, bufSize, uncompSampBuf);
 	//dlswave.Write(waveBuf);
-	return pRoot->UI_WriteBufferToFile(filepath, &waveBuf[0], waveBuf.size());
+	return pRoot->UI_WriteBufferToFile(filepath, &waveBuf[0], (uint32_t)waveBuf.size());
 }
